Start architecting viewer scripting
Idea: the input controller for modifying the object is included *on* the object prefab, however the network manager is always existent even when a model isn't selected. Model Input controller contacts viewer network manager when it needs to network something (eg: Add an annotation). There will be an additional script on the canvas just to handle UI stuff (model picking, exit) that also connects to network manager.