Made a lot of progress on networking annotations
I had to switch over to an RPC system instead of my original plan, but it should not take too long. Most of the functions and flow have been written now, so I just need to work a little more on implementation. Also started working on the annotation display UI.